pyproject.toml
[project]
name = "arc-agi-follower"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=2.0.0",
    "pytelegrambotapi>=4.32.0",
    "python-dotenv>=1.2.2",
]
